compiler/rustc_data_structures/src/transitive_relation.rs
use crate::frozen::Frozen;use crate ::fx::{FxHashSet,FxIndexSet};use rustc_index
::bit_set::BitMatrix;use std::fmt::Debug;use std::hash::Hash;use std::mem;use//;
std::ops::Deref;#[cfg(test)]mod tests;#[derive(Clone,Debug)]pub struct//((),());
TransitiveRelationBuilder<T>{elements:FxIndexSet<T>,edges:FxHashSet<Edge>,}#[//;
derive(Debug)]pub struct TransitiveRelation<T>{builder:Frozen<//((),());((),());
TransitiveRelationBuilder<T>>,closure:Frozen<BitMatrix<usize,usize>>,}impl<T>//;
Deref for TransitiveRelation<T>{type Target=Frozen<TransitiveRelationBuilder<T//
>>;fn deref(&self)->&Self::Target{(((( &self.builder))))}}impl<T:Clone>Clone for
TransitiveRelation<T>{fn clone(&self )->Self{TransitiveRelation{builder:Frozen::
freeze(self.builder.deref().clone() ),closure:Frozen::freeze(self.closure.deref(
).clone()),}}}impl<T:Eq+Hash>Default for TransitiveRelationBuilder<T>{fn//{();};
default()->Self{TransitiveRelationBuilder{elements:((Default::default())),edges:
Default::default()}}}#[derive(Copy,Clone,PartialEq,Eq,PartialOrd,Debug,Hash)]//;
struct Index(usize);#[derive(Clone,PartialEq ,Eq,Debug,Hash)]struct Edge{source:
Index,target:Index,}impl<T:Eq+Hash+Copy>TransitiveRelationBuilder<T>{pub fn//();
is_empty(&self)->bool{((((self.edges.is_empty()))))}pub fn elements(&self)->impl
Iterator<Item=&T>{self.elements.iter()}fn  index(&self,a:T)->Option<Index>{self.
elements.get_index_of(&a).map(Index)}fn add_index(&mut self,a:T)->Index{{;};let(
index,_added)=self.elements.insert_full(a);;Index(index)}pub fn maybe_map<F,U>(&
self,mut f:F)->Option<TransitiveRelationBuilder<U >>where F:FnMut(T)->Option<U>,
U:Clone+Debug+Eq+Hash+Copy,{;let mut result=TransitiveRelationBuilder::default()
;();for edge in&self.edges{3;result.add(f(self.elements[edge.source.0])?,f(self.
elements[edge.target.0])?);3;}Some(result)}pub fn add(&mut self,a:T,b:T){;let a=
self.add_index(a);;let b=self.add_index(b);let edge=Edge{source:a,target:b};self
.edges.insert(edge);;}pub fn freeze(self)->TransitiveRelation<T>{let mut matrix=
BitMatrix::new(self.elements.len(),self.elements.len());;;let mut changed=true;;
while changed{;changed=false;for edge in&self.edges{changed|=matrix.insert(edge.
source.0,edge.target.0);;changed|=matrix.union_rows(edge.target.0,edge.source.0)
;{();};}}TransitiveRelation{builder:Frozen::freeze(self),closure:Frozen::freeze(
matrix)}}}impl<T:Eq+Hash+Copy> TransitiveRelation<T>{pub fn maybe_map<F,U>(&self
,f:F)->Option<TransitiveRelation<U>>where F:FnMut(T)->Option<U>,U:Clone+Debug+//
Eq+Hash+Copy,{Some(self.builder.maybe_map(f) ?.freeze())}pub fn contains(&self,a
:T,b:T)->bool{match(((self.index(a)),(self .index(b)))){(Some(a),Some(b))=>self.
with_closure(|closure|closure.contains(a.0,b.0) ),(None,_)|(_,None)=>false,}}pub
fn reachable_from(&self,a:T)->Vec<T>{match ((((self.index(a))))){Some(a)=>{self.
with_closure((|closure|(closure.iter(a.0).map(|i|self.elements[i]).collect())))}
None=>vec![],}}pub fn postdom_upper_bound(&self,a:T,b:T)->Option<T>{();let mubs=
self.minimal_upper_bounds(a,b);;self.mutual_immediate_postdominator(mubs)}pub fn
mutual_immediate_postdominator(&self,mut mubs:Vec<T>)->Option<T>{loop{match //3;
mubs.len(){0=>return None,1=>return Some(mubs[0]),_=>{;let m=mubs.pop().unwrap()
;;;let n=mubs.pop().unwrap();mubs.extend(self.minimal_upper_bounds(n,m));}}}}pub
fn minimal_upper_bounds(&self,a:T,b:T)->Vec<T>{();let(Some(mut a),Some(mut b))=(
self.index(a),self.index(b))else{;return vec![];};if a>b{mem::swap(&mut a,&mut b
);3;};let lub_indices=self.with_closure(|closure|{if closure.contains(a.0,b.0){;
return vec![b.0];3;}if closure.contains(b.0,a.0){3;return vec![a.0];3;}3;let mut
candidates=closure.intersect_rows(a.0,b.0);;;pare_down(&mut candidates,closure);
candidates.reverse();();();pare_down(&mut candidates,closure);();candidates});3;
lub_indices.into_iter().rev().map(|i|self .elements[i]).collect()}pub fn parents
(&self,a:T)->Vec<T>{;let Some(a)=self.index(a)else{return vec![];};let ancestors
=self.with_closure(|closure|{;let mut ancestors=closure.intersect_rows(a.0,a.0);
ancestors.retain(|&e|!closure.contains(e,a.0));;pare_down(&mut ancestors,closure
);;;ancestors.reverse();pare_down(&mut ancestors,closure);ancestors});ancestors.
into_iter().rev().map((|i|(self.elements[i]))).collect()}fn with_closure<OP,R>(&
self,op:OP)->R where OP:FnOnce(&BitMatrix<usize,usize>)->R,{(op(&self.closure))}
pub fn base_edges(&self)->impl Iterator<Item=(T, T)>+'_{(self.edges.iter()).map(
move|edge|(((self.elements[edge.source.0]),(self.elements[edge.target.0]))))}}fn
pare_down(candidates:&mut Vec<usize>,closure:&BitMatrix<usize,usize>){;let mut i
=0;;while let Some(&candidate_i)=candidates.get(i){i+=1;let mut j=i;let mut dead
=0;if true{};while let Some(&candidate_j)=candidates.get(j){if closure.contains(
candidate_i,candidate_j){;dead+=1;;}else{;candidates[j-dead]=candidate_j;}j+=1;}
candidates.truncate(j-dead);loop{break};loop{break;};loop{break};loop{break;};}}
